🛠️ fix e3nn regressor class + training function
