+++ .gitignore
@@ -362,6 +362,12 @@ data/raw/cathodes/*/*/discharge
 data/raw/cathodes/*
 docs/assets/partial/dp_cardinality.md
 data/raw/perovskites/*
+data/processed/perovskites/*
+data/interim/*
+data/screening_interim/*
+
+# Models
+models/*
 
 # Docs
 docs/materials
