--- data/interim/cathodes/insertion/Na/tofinal.yaml
@@ -1,22 +1,0 @@
-path: data/final/cathodes/insertion/Na/candidates.json
-note: ""
-columns:
-  MaterialId: Material Id
-  Composition: Composition
-  Crystal System: Crystal System
-  Formation Energy Per Atom: Formation Energy (eV/atom)
-  Reduced Formula: Formula
-  average_voltage_mean: Average Voltage (V)
-  average_voltage_std: Average Voltage (deviation) (V)
-  classifier_mean: AI-experts confidence (-)
-  classifier_std: AI-experts confidence (deviation) (-)
-  max_delta_volume_mean: Max Volume expansion (-)
-  max_delta_volume_std: Max Volume expansion (deviation) (-)
-  stability_charge_mean: Stability charge (eV/atom)
-  stability_charge_std: Stability charge (deviation) (eV/atom)
-  stability_discharge_mean: Stability discharge (eV/atom)
-  stability_discharge_std: Stability discharge (deviation) (eV/atom)
-  capacity_vol: Volumetric capacity (mAh/L)
-  capacity_grav: Gravimetric capacity (mAh/g)
-  energy_vol: Volumetric energy (Wh/L)
-  energy_grav: Gravimetric energy (Wh/kg)
