--- data/interim/perovskites/pure_models/tofinal.yaml
@@ -1,14 +1,0 @@
-path: data/final/perovskites/pure_models/candidates.json
-note: ""
-columns:
-  MaterialId: Material Id
-  Composition: Composition
-  Crystal System: Crystal System
-  Formation Energy Per Atom: Formation Energy (eV/atom)
-  Reduced Formula: Formula
-  Volume: Volume (A^3)
-  Density: Density (A^3/atom)
-  band_gap_mean: Average Band Gap (eV)
-  band_gap_std: Average Band Gap (deviation) (eV)
-  classifier_mean: AI-experts confidence (-)
-  classifier_std: AI-experts confidence (deviation) (-)
