--- data/interim/thermoelectrics/690K/tofinal.yaml
@@ -1,16 +1,0 @@
-path: data/final/thermoelectrics/690K/candidates.json
-note: ""
-columns:
-  MaterialId: Material Id
-  Composition: Composition
-  Crystal System: Crystal System
-  Formation Energy Per Atom: Formation Energy (eV/atom)
-  Reduced Formula: Formula
-  Volume: Volume (A^3)
-  Density: Density (A^3/atom)
-  temperature(K): Working Temperature (K)
-  mean_regressors: Average zT (-)
-  std_regressors: Average zT (deviation) (-)
-  mean_classifiers: AI-experts confidence (-)
-  std_classifiers: AI-experts confidence (deviation) (-)
-  Notes: Notes
