+++ energy_gnome/dataset/__init__.py
@@ -3,8 +3,10 @@ from pathlib import Path  # noqa: F401
 from energy_gnome.config import DATA_DIR
 from energy_gnome.utils.logger_config import logger
 
+from .base_dataset import BaseDatabase
 from .cathodes import CathodeDatabase
 from .perovskites import PerovskiteDatabase
+from .processed_data import process_perovskite
 
 # from .thermoelectrics import ThermoelectricDatabase
 from .raw_data import get_raw_cathode, get_raw_perovskite
