+++ energy_gnome/main.py
@@ -13,12 +13,13 @@ from energy_gnome.config import (  # noqa: F401
     RAW_DATA_DIR,
     WORKING_IONS,
 )
-from energy_gnome.dataset import get_raw_cathode, get_raw_perovskite
+from energy_gnome.dataset import get_raw_cathode, get_raw_perovskite, process_perovskite
 from energy_gnome.dataset.cathodes import CathodeDatabase
 from energy_gnome.dataset.perovskites import PerovskiteDatabase
 
 # from energy_gnome.dataset.temp_convert import process_data_with_yaml
-from energy_gnome.models.regressor import E3NNRegressor, fit_regressor
+from energy_gnome.models.e3nn.regressor import E3NNRegressor
+from energy_gnome.models.train import fit_regressor
 from energy_gnome.utils.logger_config import logger, setup_logging  # noqa: F401
 
 # Create the Typer app
@@ -75,6 +76,7 @@ def datasets(
 @app.command()
 def pre_processing(
     energy_material: str = typer.Argument(..., help="Type of energy material (e.g., 'cathode')."),
+    data_dir: Path = typer.Option(DATA_DIR, help="Directory path to read raw data."),
     raw_data_dir: Path = typer.Option(DATA_DIR, help="Directory path to read raw data."),
     processed_data_dir: Path = typer.Option(
         DATA_DIR, help="Directory path to save processed data."
@@ -100,19 +102,16 @@ def pre_processing(
         for battery_type in BATTERY_TYPES:
             for working_ion in WORKING_IONS:
                 logger.info(
-                    f"--- Building regressors for working ion '{working_ion}' and '{battery_type}' battery type ".ljust(
+                    f"--- Cleaning cathode database for working ion '{working_ion}' and '{battery_type}' battery type ".ljust(
                         LOG_MAX_WIDTH, "-"
                     )
                 )
                 # TODO: implement cathode preprocessing logic
 
     elif energy_material.lower() in ["photovoltaic", "photovoltaics", "perovskites", "perovskite"]:
-        category = "perovskites"
-        logger.info("--- Building regressors for perovskites".ljust(LOG_MAX_WIDTH, "-"))
-        clean_perov_db(
-            category=category,
-            raw_data_dir=raw_data_dir,
-            processed_data_dir=processed_data_dir,
+        logger.info("--- Cleaning perovskites database".ljust(LOG_MAX_WIDTH, "-"))
+        process_perovskite(
+            data_dir=data_dir,
             logger=logger,
         )
     else:
@@ -124,7 +123,8 @@ def pre_processing(
 @app.command()
 def training(
     energy_material: str = typer.Argument(..., help="Type of energy material (e.g., 'cathode')."),
-    model_type: str = typer.Argument(..., help="Type of model to train (e.g., 'regressor')."),
+    problem_type: str = typer.Argument(..., help="Type of problem (e.g., 'regressor')."),
+    model_type: str = typer.Argument(..., help="Type of model to train (e.g., 'e3nn')."),
     data_dir: Path = typer.Option(
         PROCESSED_DATA_DIR, help="Directory path to access processed data."
     ),
@@ -145,7 +145,7 @@ def training(
     logger.info(" RAW DATASET ".center(LOG_MAX_WIDTH, "+"))
 
     if energy_material.lower() in ["cathode", "cathodes", "batteries", "battery"]:
-        if model_type.lower() in ["regressor", "regressors", "regression"]:
+        if problem_type.lower() in ["regressor", "regressors", "regression"]:
             for battery_type in BATTERY_TYPES:
                 for working_ion in WORKING_IONS:
                     logger.info(
@@ -153,7 +153,7 @@ def training(
                             LOG_MAX_WIDTH, "-"
                         )
                     )
-        elif model_type.lower() in ["classifier", "classifiers", "classification"]:
+        elif problem_type.lower() in ["classifier", "classifiers", "classification"]:
             for battery_type in BATTERY_TYPES:
                 for working_ion in WORKING_IONS:
                     logger.info(
@@ -162,20 +162,29 @@ def training(
                         )
                     )
         else:
-            raise NotImplementedError(f"'{model_type}' is not a valid model type entry.")
+            raise NotImplementedError(f"'{problem_type}' is not a valid problem type entry.")
+
     elif energy_material.lower() in ["photovoltaic", "photovoltaics", "perovskites", "perovskite"]:
         category = "perovskites"
         target_property = "band_gap"
-        if model_type.lower() in ["regressor", "regressors", "regression"]:
-            logger.info("--- Building regressors for perovskites".ljust(LOG_MAX_WIDTH, "-"))
+        db = PerovskiteDatabase(DATA_DIR)
+        db.load_all()
+
+        if problem_type.lower() in ["regressor", "regressors", "regression"]:
+            if model_type.lower() in ["e3nn"]:
+                model = E3NNRegressor()
+            logger.info(
+                f"--- Building '{model_type}' regressors for perovskites".ljust(LOG_MAX_WIDTH, "-")
+            )
             fit_regressor(
-                category=category,
-                target_property=target_property,
-                data_dir=data_dir,
-                models_dir=models_dir,
+                model,
+                db,
+                category,
+                target_property,
                 logger=logger,
             )
-        elif model_type.lower() in ["classifier", "classifiers", "classification"]:
+
+        elif problem_type.lower() in ["classifier", "classifiers", "classification"]:
             logger.info("--- Building classifiers for perovskites".ljust(LOG_MAX_WIDTH, "-"))
             fit_classifier(
                 category=category,
@@ -183,21 +192,14 @@ def training(
                 models_dir=models_dir,
                 logger=logger,
             )
+
+        else:
+            raise NotImplementedError(f"'{problem_type}' is not a valid problem type entry.")
+
     else:
         raise NotImplementedError(
             f"The database for the energy material '{energy_material}' is not supported."
         )
-
-
-# @app.command()
-# def convert(
-#     database_path: str = typer.Argument(help="Interim database to preprocess."),
-#     yaml_file: str = typer.Argument(help="Yaml configuration file."),
-# ):
-#     df, final_output_path = process_data_with_yaml(database_path, yaml_file)
-#     out_dir = Path(os.path.dirname(final_output_path))
-#     out_dir.mkdir(parents=True, exist_ok=True)
-#     df.to_json(final_output_path)
 
 
 def main():
