--- energy_gnome/models/regressor/__init__.py
@@ -1,13 +1,0 @@
-from pathlib import Path  # noqa: F401
-
-from energy_gnome.config import (
-    DEFAULT_E3NN_SETTINGS,
-    DEFAULT_E3NN_TRAINING_SETTINGS,
-    DEFAULT_OPTIM_SETTINGS,
-    MODELS_DIR,
-    PROCESSED_DATA_DIR,
-)
-from energy_gnome.utils.logger_config import logger
-
-from ..train import fit_regressor
-from .e3nn_model import E3NNRegressor
