--- energy_gnome/tools/postprocessing/__init__.py
@@ -1,1 +1,0 @@
-from .dataset_statistics import get_neighbors
