--- models/perovskites/model.band_gap.250121.rep2.json
@@ -1,1 +1,0 @@
-{"in_dim": 118, "em_dim": 64, "irreps_in": "64x0e", "irreps_out": "1x0e", "irreps_node_attr": "64x0e", "layers": 2, "mul": 32, "lmax": 2, "max_radius": 5.0, "num_neighbors": 35.894536932277354, "reduce_output": true}
